Support WS and HTTP for ethereum server
