+++ chains/ethereum/server/Cargo.toml
@@ -11,13 +11,14 @@ anyhow = "1.0.69"
 async-std = { version = "1.12.0", features = ["tokio1"] }
 async-trait = "0.1.66"
 ethabi = "18.0.0"
-ethers = "2.0.2"
+ethers = { version = "2.0.9", default-features = true, features = ["ws", "abigen", "rustls"] }
 hex = "0.4.3"
 rosetta-config-ethereum = { version = "0.4.0", path = "../config" }
 rosetta-server = { version = "0.4.0", path = "../../../rosetta-server" }
 serde = "1.0.153"
 serde_json = "1.0"
 tokio = { version = "1.26.0", features = ["rt-multi-thread", "macros"] }
+url = "2.4.0"
 
 [dev-dependencies]
 ethers-solc = "2.0.1"
