+++ chains/ethereum/server/src/main.rs
@@ -2,5 +2,5 @@ use anyhow::Result;
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    rosetta_server::main::<rosetta_server_ethereum::EthereumClient>().await
+    rosetta_server::main::<rosetta_server_ethereum::GenericEthereumClient>().await
 }
